Cleaned code for submission

+++ remove_wrong_sized.py
@@ -8,8 +8,11 @@ import generate_dataset
 BASE_DIR = utils.BASE_DIR
 DIM = generate_dataset.DIM
 
-def remove(tipo: str):
-    image_path = BASE_DIR / f"ds_{tipo}_images"
+def remove_wrong(tipo: str):
+    if tipo == "test":
+        image_path = BASE_DIR / "test_image"
+    else:
+        image_path = BASE_DIR / f"ds_{tipo}_images"
     l = os.listdir(image_path)
     print(len(l))
 
@@ -17,12 +20,12 @@ def remove(tipo: str):
         img = l[i]
         img_region = rasterio.open(image_path / img)
         img_array = img_region.read()
-        #if img_array.max() != 0:
-        #    break
         if img_array.shape != (3, DIM, DIM):
             os.remove(image_path / img)
-            os.remove(BASE_DIR / f"ds_{tipo}_label" / img)
+            if tipo != "test":
+                os.remove(BASE_DIR / f"ds_{tipo}_label" / img)
 
 if __name__ == "__main__":
-    remove("train")
-    remove("val")
+    remove_wrong("train")
+    remove_wrong("val")
+    remove_wrong("test")
